emulator: Add a comment justifying storing the column family iterator in the undo log.
In general, it is unsafe to store other kind of map iterators in the
undo log, since a subsequent mutation on the same row could delete the
underlying map, and invalidate the iterator. However, deleting a
column family is a schema operation that cannot happen while we hold
the table lock, therefore for column families only, it is safe to
store the iterators in the undo log.