emulator: RowTransaction: replace request data member with row_key.
We only ever use request.row_key() when we use request in the
RowTransaction class. So we need only construct it with the row_key.

This paves the way for some refactoring that would allow us to
execute an array of mutations (with rollback on error) without first
creating a bigtable MutateRowRequest.