fix: fix segmentation fault due to the use of an invalid grpc server_ object.
BuildAndStart() can fail for any number of reasons (e.g. when passed a
port reserved for privileged use or a non-existent local address). We
were not checking for this and would try to use the server
unconditionally, causing a segmentation fault (e.g. when emulator was
called by a non-root user as: ./emulator -p 100).

This adds a fix so that we can check that we have a valid server
object before trying to use it.